requirements.txt
django
pandas
openpyxl
django-import-export